Added a logger to the action (#5)
Using the npm logger instead of the console logger since the console loggeer was designed to run in a browser and not node. It is not performant and runs into concurrency issues, so I decided to bite the bullet and use the npm logger sooner rather than later.

I chose the npm logger since it appears to be popular and very simple to use. For such a simple project, it is easier to pick a popular logger and move on :)