PR 11 Fast follow (#12)
* Only deploy action output on push to main
Until today every time the action "e2e test" was run it would push the file to the S3 bucket. The purpose of that file is to ensure the action works correctly, not that the deploying to s3 works.

I believe this is the right time to only push updated files to s3 on pushes to main since I have manually verified that the flow works and the latex file is rendered correctly. My assumption going forward will be that any issues with the compilation of the latex files will fail the build.

* Forgot to commit the change where the deploy action only runs on pushes to main 🤦